Kullanici Converter Olusturuldu gerekli duzenleme yapildi

+++ src/java/dao/KullaniciDAO.java
@@ -4,7 +4,6 @@ import entity.Kullanici;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -48,7 +47,7 @@ public class KullaniciDAO extends Dao {
         return count;
     }
 
-    public Kullanici find(int kullaniciid) {
+    public Kullanici find(Long kullaniciid) {
         Kullanici k = null;
 
         try {
